src/main/java/dad/battle/character/Character.java
package dad.battle.character;

import dad.battle.die.Dice;

public class Character {

	private static final int STARTING_LEVEL = 1;
	private String name;
	private int strength;
	private int dexterity;
	private int constitution;
	private int intelligence;
	private int wisdom;
	private int charisma;
	private int hitPoints;
	private int level;
	private Classs classs;

	public Character(String name) {
		super();
		this.name = name;
		Dice abilityDice = new Dice("3d6");
		strength = abilityDice.roll();
		dexterity = abilityDice.roll();
		constitution = abilityDice.roll();
		intelligence = abilityDice.roll();
		wisdom = abilityDice.roll();
		charisma = abilityDice.roll();
		level = STARTING_LEVEL;
		classs = Classs.FIGHTER;
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public int getStrength() {
		return strength;
	}

	public void setStrength(int strength) {
		this.strength = strength;
	}

	public int getDexterity() {
		return dexterity;
	}

	public void setDexterity(int dexterity) {
		this.dexterity = dexterity;
	}

	public int getConstitution() {
		return constitution;
	}

	public void setConstitution(int constitution) {
		this.constitution = constitution;
	}

	public int getIntelligence() {
		return intelligence;
	}

	public void setIntelligence(int intelligence) {
		this.intelligence = intelligence;
	}

	public int getWisdom() {
		return wisdom;
	}

	public void setWisdom(int wisdom) {
		this.wisdom = wisdom;
	}

	public int getCharisma() {
		return charisma;
	}

	public void setCharisma(int charisma) {
		this.charisma = charisma;
	}

	public int getHitPoints() {
		return hitPoints;
	}

	public void setHitPoints(int hitPoints) {
		this.hitPoints = hitPoints;
	}

	public int getLevel() {
		return level;
	}

	public void setLevel(int level) {
		this.level = level;
	}

	public Classs getClasss() {
		return classs;
	}

	public void setClasss(Classs classs) {
		this.classs = classs;
	}

}
